+++ src/inc/drawer.js
@@ -1,0 +1,78 @@
+import React from "react";
+import Drawer from "@material-ui/core/Drawer";
+import Toolbar from "@material-ui/core/Toolbar";
+import List from "@material-ui/core/List";
+import ListItem from "@material-ui/core/ListItem";
+import ListItemIcon from "@material-ui/core/ListItemIcon";
+import InboxIcon from "@material-ui/icons/MoveToInbox";
+import MailIcon from "@material-ui/icons/Mail";
+import ListItemText from "@material-ui/core/ListItemText";
+import Divider from "@material-ui/core/Divider";
+import {makeStyles} from "@material-ui/core/styles";
+import {Link} from "react-router-dom";
+
+const drawerWidth = 240;
+
+const useStyles = makeStyles((theme) => ({
+    drawer: {
+        width: drawerWidth,
+        flexShrink: 0,
+    },
+    drawerPaper: {
+        width: drawerWidth,
+    },
+    drawerContainer: {
+        color: '#000',
+        overflow: 'auto',
+    },
+    content: {
+        flexGrow: 1,
+        padding: theme.spacing(3),
+    },
+    link: {
+        color: '#000',
+        textDecoration: 'none',
+    },
+}));
+
+const _Drawer = () => {
+    const classes = useStyles();
+
+    const path = (text) => {
+        return (text === '일기장' ? '/diary' : text === '멘토링' ? '/mentoring' : '/board')
+    }
+
+    return (
+        <Drawer
+            className={classes.drawer}
+            variant="permanent"
+            classes={{
+                paper: classes.drawerPaper,
+            }}
+        >
+            <Toolbar/>
+            <div className={classes.drawerContainer}>
+                <List>
+                    {['일기장', '멘토링', '게시판'].map((text, index) => (
+                        <Link className={classes.link} to={path(text)}>
+                            <ListItem button key={text}>
+                                <ListItemIcon>{index % 2 === 0 ? <InboxIcon/> : <MailIcon/>}</ListItemIcon>
+                                <ListItemText primary={text}/>
+                            </ListItem>
+                        </Link>
+                    ))}
+                </List>
+                <Divider/>
+                <List>
+                    {['About'].map((text, index) => (
+                        <ListItem button key={text}>
+                            <ListItemIcon>{index % 2 === 0 ? <InboxIcon/> : <MailIcon/>}</ListItemIcon>
+                            <ListItemText primary={text}/>
+                        </ListItem>
+                    ))}
+                </List>
+            </div>
+        </Drawer>
+    );
+}
+export default _Drawer;
