+++ src/inc/header.js
@@ -1,0 +1,26 @@
+import React from 'react';
+import {makeStyles} from "@material-ui/core/styles";
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
+
+const useStyles = makeStyles((theme) => ({
+    appBar: {
+        zIndex: theme.zIndex.drawer + 1,
+    },
+}));
+
+const Head = () => {
+    const classes = useStyles();
+    return (
+        <AppBar position="fixed" className={classes.appBar}>
+            <Toolbar>
+                <Typography variant="h6" noWrap>
+                    스터디로그
+                </Typography>
+            </Toolbar>
+        </AppBar>
+    );
+}
+
+export default Head;
