+++ src/inc/index.js
@@ -1,0 +1,2 @@
+export {default as _Drawer} from './drawer';
+export {default as Head} from './header';
