+++ src/pages/about.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const About = () => {
+
+    return (
+        <div>
+
+        </div>
+    );
+}
+export default About;
