+++ src/pages/index.js
@@ -1,4 +1,6 @@
 export {default as Login} from './login';
-export {default as Home} from './home';
 export {default as Diary} from './diary';
 export {default as Write} from './write';
+export {default as About} from './about';
+export {default as Main} from './main';
+export {default as Mentoring} from './mentoring';
