+++ src/pages/mentoring.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Mentoring = () => {
+
+    return (
+        <div>
+
+        </div>
+    );
+}
+export default Mentoring;
